requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# HTTP requests
requests==2.31.0

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Environment variables
python-dotenv==1.0.0

# Google Gemini
google-generativeai==0.3.2